changed entities
