+++ common/pom.xml
@@ -17,18 +17,12 @@
     <spring.version>5.3.14</spring.version>
   </properties>
   <packaging>jar</packaging>
-
   <dependencies>
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>3.8.1</version>
       <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>by.tsarenkov</groupId>
-      <artifactId>service</artifactId>
-      <version>1</version>
     </dependency>
   </dependencies>
 </project>
