+++ common/src/main/java/by/tsarenkov/common/model/entity/Author.java
@@ -17,9 +17,9 @@ import java.util.Collection;
 @EqualsAndHashCode
 public class Author  implements Serializable {
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id_author")
-    private long id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_author", unique = true)
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "surname")
@@ -31,7 +31,7 @@ public class Author  implements Serializable {
     @ManyToMany
     @JoinTable(name = "author_book",
             joinColumns = @JoinColumn(name = "id_author"),
-            inverseJoinColumns = @JoinColumn(name = "id_author")
+            inverseJoinColumns = @JoinColumn(name = "id_book")
     )
     private Collection<Book> books;
 }
