+++ db/src/main/java/by/tsarenkov/db/AuthorRepository.java
@@ -1,0 +1,9 @@
+package by.tsarenkov.db;
+
+import by.tsarenkov.common.model.entity.Author;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AuthorRepository extends CrudRepository<Author, Long> {
+}
