--- db/src/main/java/by/tsarenkov/db/repository/AuthorRepository.java
@@ -1,7 +1,0 @@
-package by.tsarenkov.db.repository;
-
-import by.tsarenkov.common.model.entity.Author;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AuthorRepository extends JpaRepository<Author, Long> {
-}
