+++ service/src/main/java/by/tsarenkov/service/AuthorService.java
@@ -1,7 +1,9 @@
 package by.tsarenkov.service;
 
 import by.tsarenkov.common.model.entity.Author;
+import org.springframework.stereotype.Service;
 
+@Service
 public interface AuthorService {
     void saveAuthor(Author author);
     void deleteAuthor(Author author);
