+++ service/src/test/java/AuthorServiceTest.java
@@ -7,18 +7,4 @@ import org.junit.Test;
 
 public class AuthorServiceTest {
 
-    static AuthorService authorService = new AuthorServiceImpl();
-    static Author author;
-
-    @BeforeClass
-    public static void initList() {
-        author = Author.builder().name("Александр")
-                .patronymic("Сергеевич")
-                .surname("Пушкин").build();
-    }
-
-    @Test
-    public void saveAuthor() {
-        //authorService.saveAuthor(author);
-    }
 }
